backend/requirements.txt
django-cors-headers==3.11.0
Django==4.0
djangorestframework==3.13.1
gunicorn==20.0.4
psycopg2-binary==2.8.6